+++ posthog/api/test/test_authentication.py
@@ -17,34 +17,12 @@ from posthog.test.base import APIBaseTest
 
 
 class TestLoginPrecheckAPI(APIBaseTest):
+    """
+    Tests the login precheck API.
+    Please note additional login tests are included in ee/api/test/test_authentication.py
+    """
+
     CONFIG_AUTO_LOGIN = False
-
-    def test_login_precheck_with_enforced_sso(self):
-        OrganizationDomain.objects.create(
-            domain="witw.app",
-            organization=self.organization,
-            verified_at=timezone.now(),
-            sso_enforcement="google-oauth2",
-        )
-        User.objects.create_and_join(self.organization, "spain@witw.app", self.CONFIG_PASSWORD)
-
-        response = self.client.post("/api/login/precheck", {"email": "spain@witw.app"})
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.json(), {"sso_enforcement": "google-oauth2"})
-
-    def test_login_precheck_with_unverified_domain(self):
-        OrganizationDomain.objects.create(
-            domain="witw.app",
-            organization=self.organization,
-            verified_at=None,  # note domain is not verified
-            sso_enforcement="google-oauth2",
-        )
-
-        response = self.client.post(
-            "/api/login/precheck", {"email": "i_do_not_exist@witw.app"}
-        )  # Note we didn't create a user that matches, only domain is matched
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.json(), {"sso_enforcement": None})
 
     def test_login_precheck_with_unenforced_sso(self):
         OrganizationDomain.objects.create(
@@ -55,21 +33,17 @@ class TestLoginPrecheckAPI(APIBaseTest):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.json(), {"sso_enforcement": None})
 
-    def test_login_precheck_with_inexistent_account(self):
-        OrganizationDomain.objects.create(
-            domain="anotherdomain.com",
-            organization=self.organization,
-            verified_at=timezone.now(),
-            sso_enforcement="github",
-        )
-        User.objects.create_and_join(self.organization, "i_do_not_exist@anotherdomain.com", self.CONFIG_PASSWORD)
-
-        response = self.client.post("/api/login/precheck", {"email": "i_do_not_exist@anotherdomain.com"})
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.json(), {"sso_enforcement": "github"})
+    def test_login_precheck_with_sso_enforced_with_invalid_license(self):
+        # TODO
+        pass
 
 
 class TestLoginAPI(APIBaseTest):
+    """
+    Tests the general password login API.
+    Please note additional login tests are included in ee/api/test/test_authentication.py (e.g. testing SSO enforcement)
+    """
+
     CONFIG_AUTO_LOGIN = False
 
     @patch("posthoganalytics.capture")
@@ -288,13 +262,14 @@ class TestPasswordResetAPI(APIBaseTest):
             SSO_ENFORCEMENT="saml",
         ):
             response = self.client.post("/api/reset/", {"email": "i_dont_exist@posthog.com"})
+
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertEqual(
             response.json(),
             {
                 "type": "validation_error",
                 "code": "sso_enforced",
-                "detail": "Password reset is disabled because SSO login is enforced.",
+                "detail": "Password reset is disabled because SSO login is enforced for this instance.",
                 "attr": None,
             },
         )
