+++ posthog/utils.py
@@ -687,27 +687,6 @@ def get_available_sso_providers() -> Dict[str, bool]:
     return output
 
 
-def get_sso_enforced_provider() -> Optional[str]:
-    """
-    Returns the enforced SSO provider handle for the instance if SSO is properly configured and required license is present.
-        => response: `saml`, `google-oaut2`, `github`, `gitlab`, `None`.
-    """
-    sso_enforcement = getattr(settings, "SSO_ENFORCEMENT", None)
-
-    if sso_enforcement:
-        sso_providers = get_available_sso_providers()
-        if not sso_providers[settings.SSO_ENFORCEMENT]:
-            print_warning(
-                [
-                    f"You have configured `SSO_ENFORCEMENT` with value `{settings.SSO_ENFORCEMENT}`,"
-                    " but that provider is not properly configured or your instance does not have the required license."
-                ]
-            )
-            return None
-
-    return sso_enforcement
-
-
 def flatten(i: Union[List, Tuple]) -> Generator:
     for el in i:
         if isinstance(el, list):
